+++ src/component/Gem/Gem.js
@@ -1,9 +1,10 @@
 import React, { useState, useContext } from 'react'
+import { AppC } from '../Context';
 import { bigData } from '../myContext';
 
 function Gem() {
   const [tip, setTip] = useState([false,false,false,false,false,false,false,false,false,false,false]);
-  const { gem, effects, legend, hero } = useContext(bigData)
+  const {gem, effects, legend, hero} = useContext(AppC);
 
   //보석 마우스 아웃 시 이벤트
   const mouseOut = () => {
