+++ src/component/Insert/Insert.js
@@ -1,8 +1,9 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import { Link } from 'react-router-dom'
+import { AppC } from '../Context'
 
-function Insert( { search, elName } ) {
-
+function Insert( ) {
+  const {elName, search} = useContext(AppC);
   return (
     <>
     <div>
